+++ src/components/Plugin/PluginType.js
@@ -1,0 +1,5 @@
+const PluginType = {
+  CreateToken: "CreateToken",
+  Tip: "Tip",
+};
+export default PluginType;