+++ CHANGELOG.md
@@ -165,7 +165,7 @@ ajv                        6.9.1   →   6.10.0
 - **src/components/Tasks/Tasks.jsx**
 - **src/views/Notifications/Notifications.jsx**
 ### New components
-- **src/components/FixedPlugin/FixedPlugin.jsx**(Left menu used by us for customization - due to this, we've had to add a link for the fonts awesome cdn)
+- **src/components/CreateTokenPlugin/CreateTokenPlugin.jsx**(Left menu used by us for customization - due to this, we've had to add a link for the fonts awesome cdn)
 - **src/components/Header/RTLHeaderLinks.jsx**
 - **src/layouts/Admin.jsx** (instead of Dashboard)
 - **src/layouts/RTL.jsx**